add headline field item
