+++ heartbeat-app/src/main/kotlin/edu/stanford/bdh/heartbeat/app/survey/ui/QuestionNumberInfo.kt
@@ -12,6 +12,7 @@ data class QuestionNumberInfo(
     private val current: Int,
     private val total: Int
 ) : SurveyItem {
+
     @Composable
     override fun Content(modifier: Modifier) {
         Text(
