+++ heartbeat-app/src/main/kotlin/edu/stanford/bdh/heartbeat/app/survey/ui/TextFieldItem.kt
@@ -1,16 +1,15 @@
 package edu.stanford.bdh.heartbeat.app.survey.ui
 
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import edu.stanford.bdh.heartbeat.app.survey.ui.fields.FormFieldItem
 import edu.stanford.spezi.core.design.theme.ThemePreviews
 
 data class TextFieldItem(
     override val id: String,
-    override val required: Boolean,
     val info: QuestionNumberInfo,
     val fieldLabel: QuestionFieldLabel,
     val value: String,
@@ -20,16 +19,14 @@ data class TextFieldItem(
     @Composable
     override fun Content(modifier: Modifier) {
         SurveyCard {
-            Column {
-                info.Content(Modifier)
-                fieldLabel.Content(Modifier)
-                TextField(
-                    value = value,
-                    modifier = Modifier.fillMaxWidth(),
-                    onValueChange = { onValueChange(it) },
-                    placeholder = { Text("Enter your input here...") }
-                )
-            }
+            info.Content(Modifier)
+            fieldLabel.Content(Modifier)
+            TextField(
+                value = value,
+                modifier = Modifier.fillMaxWidth(),
+                onValueChange = { onValueChange(it) },
+                placeholder = { Text("Enter your input here...") }
+            )
         }
     }
 }
@@ -39,13 +36,12 @@ data class TextFieldItem(
 private fun Previews() {
     val textFieldItem = TextFieldItem(
         id = "",
-        required = true,
         info = QuestionNumberInfo(2, 11),
         fieldLabel = QuestionFieldLabel("Phone number"),
         value = "",
         onValueChange = {},
     )
-    SurveyPreview {
+    SurveyItemPreview {
         textFieldItem.Content(Modifier)
     }
 }
