+++ heartbeat-app/src/main/kotlin/edu/stanford/bdh/heartbeat/app/survey/ui/fields/FormFieldItem.kt
@@ -1,0 +1,7 @@
+package edu.stanford.bdh.heartbeat.app.survey.ui.fields
+
+import edu.stanford.bdh.heartbeat.app.survey.ui.SurveyItem
+
+interface FormFieldItem : SurveyItem {
+    val id: String
+}
